refactor(kubernetes): add @Override annotation to Lombok-generated getters
The primary goal of this PR is to turn on error-prone's missing @Overrides warnings. This commit is additive in that it adds redundant @Getter annotations to all fields that require the @Override annotation, except in the case of class-level @Getter annotations, in which case we need to remove those in favor of field-level @Getters so that the onMethod parameter for the @Override getters takes precedence.