refactor(kubernetes): remove class-level Lombok annotations from OperationResult and KubernetesRunJobDeploymentResult
The remaining missing @Overrides warnings are for the `canEqual` methods generated by Lombok's @EqualsAndHashCode. This is a [known issue](https://github.com/rzwitserloot/lombok/issues/1343), but before we suppress the remaining warnings, let's remove unnecessary uses of @EqualsAndHashCode.

OperationResult is annotated with @Data, which is equivalent to the following annotations: @Getter, @Setter, @RequiredArgsConstructor, @ToString, and @EqualsAndHashCode. This commit replaces OperationResult's @Data and @NoArgsConstructor with field-level @Getters and @Setters and an explicit no-args constructor. As a result, KubernetesRunJobDeploymentResult no longer requires its own @EqualsAndHashCode(callSuper = true). Also replaces its @Data with field-level @Getter and @Setter. Let's try to avoid using @Data in the future for classes without a notion of logical equality that would require @EqualsAndHashCode, and instead just add the constructors and @Getters and @Setters we actually need.